Move resolve into commands
