+++ dotpkg/commands.py
@@ -1,11 +1,11 @@
 from pathlib import Path
 from typing import Iterable, Optional
 
+from dotpkg.constants import IGNORED_NAMES
 from dotpkg.install import install, uninstall
 from dotpkg.manifest.dotpkg import DotpkgManifest
 from dotpkg.model import DotpkgRef, DotpkgRefs
 from dotpkg.options import Options
-from dotpkg.resolve import cwd_dotpkgs
 from dotpkg.utils.log import info, warn
 from dotpkg.utils.prompt import confirm
 
@@ -31,6 +31,28 @@ def batch_skip_reason(manifest: DotpkgManifest) -> Optional[str]:
         return f"Could not find {', '.join(unsatisfied_reqs)} on PATH"
 
     return None
+
+def cwd_dotpkgs(opts: Options) -> DotpkgRefs:
+    cwd = opts.cwd.resolve()
+
+    # Prefer current directory if it contains a manifest
+    cwd_ref = DotpkgRef(cwd)
+    if cwd_ref.manifest_path.exists():
+        return DotpkgRefs(
+            refs=[cwd_ref],
+            is_batch=False,
+        )
+    
+    # Otherwise resolve child directories
+    return DotpkgRefs(
+        refs=[
+            ref
+            for p in cwd.iterdir()
+            for ref in [DotpkgRef(p)]
+            if not p.name in IGNORED_NAMES and ref.manifest_path.exists()
+        ],
+        is_batch=True,
+    )
 
 def resolve_refs(raw_dotpkg_paths: list[str], opts: Options) -> DotpkgRefs:
     if raw_dotpkg_paths:
